misc/HtmlTestSuite/results/test-1133.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
<meta content="script-src 'self' 'nonce-abc'" http-equiv="content-security-policy"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head>
<body>
<iframe src="support/frame-with-csp.sub.html"/>
<div id="special_div" onclick="frames[0].location.href = 'javascript:parent.postMessage(\'executed\', \'*\')'"/>
<script nonce="abc">
  var t = async_test("Should not have executed the javascript url");
  frames[0].addEventListener('load', () =&gt; {
    window.onmessage = t.step_func(function(e) {
      if (e.data == "executed")
        assert_true(false, "Javascript url executed");
    });
    window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
      assert_equals(e.blockedURI, 'inline');
      assert_equals(e.violatedDirective, 'script-src-attr');
    }));
    document.getElementById('special_div').click();
  });
</script>
</body>
</html>
